refactor: add helper for creating monthly report and moving logic from mutation
